Vb 2341 move admin endpoints into correct code base/package (#441)
* VB-2341 Move admin endpoints into correct code base / package

* VB-2341 Move admin endpoints into correct code base / package
